requirements.txt
# linebot
line-bot-sdk
Flask
# access google sheets
gspread
google-auth
# data processing
pandas
matplotlib
dataframe_image
# upload image
pyimgur
# scheduling
APScheduler